requirements.txt
streamlit
streamlit-lottie
python-dotenv
oci
langchain
faiss-cpu